+++ Console/Command/ProductImportCommand.php
@@ -82,7 +82,7 @@ class ProductImportCommand extends Command
             // fail gracefully
         }
 
-        $this->registry->register('isSecureArea', true);
+        $this->registry->register('isSecureArea', true, true);
 
         $start = $this->getCurrentMs();